requirements.txt
pandas 
openpyxl
psycopg2-binary==2.9.9